[IMP] minor.

+++ bench/run_fugaku.sh
@@ -33,16 +33,15 @@ for p in "${NODES[@]}"
 do
     echo ${SYSTEM}${EXTRA},${p},${OUT_FOLDER} >> ../data/description.csv
     
-    #for n in 1 8 64 512 2048 16384 131072 1048576 8388608 67108864
-    for n in 8 64 512 2048 16384 131072 1048576 8388608 67108864
+    for n in 1 8 64 512 2048 16384 131072 1048576 8388608 67108864
     do
         iterations=0
         if [ $n -le 512 ]
         then
-            iterations=10000
+            iterations=100 #00
         elif [ $n -le 1048576 ]
         then
-            iterations=1000
+            iterations=100 #0
         elif [ $n -le 8388608 ]
         then
             iterations=100
@@ -57,52 +56,78 @@ do
         coll_tuned_prealloc_size=1024
 
         DEFAULT_ALGO="default"
-	    EXTRA_MCAS="-mca mpi_print_stats 1 -mca coll_select_show_decision_process 2" #"-mca coll_base_reduce_commute_safe 1"
+	    # ATTENTION: Showing decision process adds non-negligible overhead (for small vectors). Use it with care.
+	    EXTRA_MCAS="" #"-mca mpi_print_stats 1 -mca coll_select_show_decision_process 2" #"-mca coll_base_reduce_commute_safe 1"
+        # TODO: Maybe I should prealloc only for large allreduce?
         LIBSWING_ALGO="DEFAULT" ${MPIRUN} ${EXTRA_MCAS} -mca coll_tuned_prealloc_size ${coll_tuned_prealloc_size} ${MPIRUN_MAP_BY_NODE_FLAG} ${MPIEXEC_OUT} -n ${p} ${MPIRUN_ADDITIONAL_FLAGS} ./bench ${COLLECTIVE} INT ${n} ${iterations}
         mv ${OUT_PREFIX}*.0 ${OUT_FOLDER}/${p}_${n}_default_${DEFAULT_ALGO}.csv; rm -f ${OUT_PREFIX}* #${ERR_PREFIX}*
         mkdir ${OUT_FOLDER}/${p}_${n}_default_${DEFAULT_ALGO}_stats/
         mv ${ERR_PREFIX}* ${OUT_FOLDER}/${p}_${n}_default_${DEFAULT_ALGO}_stats/
         
-        '''
-        for DEFAULT_ALGO in "rdbc" "trix6" "trix3" "segmented_ring" "ring" "recursive_doubling" "nonoverlapping" "basic_linear"
+        # Decision rules can be found at /opt/FJSVxtclanga/.common/MECA030/etc/fjmpi-dectree.conf (but you can access it only on computing nodes after doing pjsub)
+
+        # All those that do not have segsize as parameter
+        for DEFAULT_ALGO in "rdbc" "ring" "recursive_doubling" "nonoverlapping" "basic_linear"
         do
             LIBSWING_ALGO="DEFAULT" ${MPIRUN} ${EXTRA_MCAS} -mca coll_tuned_prealloc_size ${coll_tuned_prealloc_size} -mca coll_select_allreduce_algorithm ${DEFAULT_ALGO} ${MPIRUN_MAP_BY_NODE_FLAG} ${MPIEXEC_OUT} -n ${p} ${MPIRUN_ADDITIONAL_FLAGS} ./bench ${COLLECTIVE} INT ${n} ${iterations}
             mv ${OUT_PREFIX}*.0 ${OUT_FOLDER}/${p}_${n}_default_${DEFAULT_ALGO}.csv; rm -f ${OUT_PREFIX}* #${ERR_PREFIX}*
             mkdir ${OUT_FOLDER}/${p}_${n}_default_${DEFAULT_ALGO}_stats/
             mv ${ERR_PREFIX}* ${OUT_FOLDER}/${p}_${n}_default_${DEFAULT_ALGO}_stats/
         done
-        '''
 
-        # Run bandwidth lat optimal swing
-        LIBSWING_DIMENSIONS=${DIMENSIONS} LIBSWING_ALGO="SWING_L" ${MPIRUN} ${MPIRUN_MAP_BY_NODE_FLAG} ${MPIEXEC_OUT} -n ${p} ${MPIRUN_ADDITIONAL_FLAGS} ./bench ${COLLECTIVE} INT ${n} ${iterations}
-        mv ${OUT_PREFIX}*.0 ${OUT_FOLDER}/${p}_${n}_lat.csv; rm -f ${OUT_PREFIX}* #${ERR_PREFIX}*
-        mkdir ${OUT_FOLDER}/${p}_${n}_lat_stats/
-        mv ${ERR_PREFIX}* ${OUT_FOLDER}/${p}_${n}_lat_stats/
-	
-        for PORTS in 1 3 6
+        # Those that have segsize as parameter
+        for DEFAULT_ALGO in "trix6" "trix3" "segmented_ring"
         do
-	    MIN_ELEMS=$((PORTS * p))
-	    if [ "$n" -ge "$MIN_ELEMS" ]; then
-		#LIBSWING_DIMENSIONS=${DIMENSIONS} LIBSWING_NUM_PORTS=${PORTS} LIBSWING_ALGO="SWING_B" ${MPIRUN} ${MPIRUN_MAP_BY_NODE_FLAG} ${MPIEXEC_OUT} -n ${p} ${MPIRUN_ADDITIONAL_FLAGS} ./bench ${COLLECTIVE} INT ${n} ${iterations}
-		#mv ${OUT_PREFIX}*.0 ${OUT_FOLDER}/${p}_${n}_bw_${PORTS}_ports.csv; rm -f ${OUT_PREFIX}* #${ERR_PREFIX}*
-		#mkdir ${OUT_FOLDER}/${p}_${n}_bw_${PORTS}_ports_stats/
-		#mv ${ERR_PREFIX}* ${OUT_FOLDER}/${p}_${n}_bw_${PORTS}_ports_stats/	
+            for coll_select_allreduce_algorithm_segmentsize in 4096 16384 65536
+            do
+                LIBSWING_ALGO="DEFAULT" ${MPIRUN} ${EXTRA_MCAS} -mca coll_select_allreduce_algorithm_segmentsize ${coll_select_allreduce_algorithm_segmentsize} -mca coll_tuned_prealloc_size ${coll_tuned_prealloc_size} -mca coll_select_allreduce_algorithm ${DEFAULT_ALGO} ${MPIRUN_MAP_BY_NODE_FLAG} ${MPIEXEC_OUT} -n ${p} ${MPIRUN_ADDITIONAL_FLAGS} ./bench ${COLLECTIVE} INT ${n} ${iterations}
+                mv ${OUT_PREFIX}*.0 ${OUT_FOLDER}/${p}_${n}_default_${DEFAULT_ALGO}_${coll_select_allreduce_algorithm_segmentsize}.csv; rm -f ${OUT_PREFIX}* #${ERR_PREFIX}*
+                mkdir ${OUT_FOLDER}/${p}_${n}_default_${DEFAULT_ALGO}_${coll_select_allreduce_algorithm_segmentsize}_stats/
+                mv ${ERR_PREFIX}* ${OUT_FOLDER}/${p}_${n}_default_${DEFAULT_ALGO}_${coll_select_allreduce_algorithm_segmentsize}_stats/
+            done
+        done
+
         
-		#LIBSWING_DIMENSIONS=${DIMENSIONS} LIBSWING_NUM_PORTS=${PORTS} LIBSWING_ALGO="SWING_B_COALESCE" ${MPIRUN} ${MPIRUN_MAP_BY_NODE_FLAG} ${MPIEXEC_OUT} -n ${p} ${MPIRUN_ADDITIONAL_FLAGS} ./bench ${COLLECTIVE} INT ${n} ${iterations}
-		#mv ${OUT_PREFIX}*.0 ${OUT_FOLDER}/${p}_${n}_bw_coalesce_${PORTS}_ports.csv; rm -f ${OUT_PREFIX}* #${ERR_PREFIX}*
-		#mkdir ${OUT_FOLDER}/${p}_${n}_bw_coalesce_${PORTS}_ports_stats/
-		#mv ${ERR_PREFIX}* ${OUT_FOLDER}/${p}_${n}_bw_coalesce_${PORTS}_ports_stats/
-        
-		#LIBSWING_DIMENSIONS=${DIMENSIONS} LIBSWING_NUM_PORTS=${PORTS} LIBSWING_ALGO="SWING_B_CONT" ${MPIRUN} ${MPIRUN_MAP_BY_NODE_FLAG} ${MPIEXEC_OUT} -n ${p} ${MPIRUN_ADDITIONAL_FLAGS} ./bench ${COLLECTIVE} INT ${n} ${iterations} 
-		#mv ${OUT_PREFIX}*.0 ${OUT_FOLDER}/${p}_${n}_bw_cont_${PORTS}_ports.csv; rm -f ${OUT_PREFIX}* #${ERR_PREFIX}*
-		#mkdir ${OUT_FOLDER}/${p}_${n}_bw_cont_${PORTS}_ports_stats/
-		#mv ${ERR_PREFIX}* ${OUT_FOLDER}/${p}_${n}_bw_cont_${PORTS}_ports_stats/	
-        
-		LIBSWING_DIMENSIONS=${DIMENSIONS} LIBSWING_NUM_PORTS=${PORTS} LIBSWING_ALGO="SWING_B_UTOFU" ${MPIRUN} ${MPIRUN_MAP_BY_NODE_FLAG} ${MPIEXEC_OUT} -n ${p} ${MPIRUN_ADDITIONAL_FLAGS} ./bench ${COLLECTIVE} INT ${n} ${iterations} 
-		mv ${OUT_PREFIX}*.0 ${OUT_FOLDER}/${p}_${n}_bw_utofu_${PORTS}_ports.csv; rm -f ${OUT_PREFIX}* #${ERR_PREFIX}*
-		mkdir ${OUT_FOLDER}/${p}_${n}_bw_utofu_${PORTS}_ports_stats/
-		mv ${ERR_PREFIX}* ${OUT_FOLDER}/${p}_${n}_bw_utofu_${PORTS}_ports_stats/
+
+
+        # Run lat optimal swing
+        if [ "$n" -le 1048576 ]; then
+            LIBSWING_DIMENSIONS=${DIMENSIONS} LIBSWING_ALGO="SWING_L" ${MPIRUN} ${MPIRUN_MAP_BY_NODE_FLAG} ${MPIEXEC_OUT} -n ${p} ${MPIRUN_ADDITIONAL_FLAGS} ./bench ${COLLECTIVE} INT ${n} ${iterations}
+            mv ${OUT_PREFIX}*.0 ${OUT_FOLDER}/${p}_${n}_lat.csv; rm -f ${OUT_PREFIX}* #${ERR_PREFIX}*
+            mkdir ${OUT_FOLDER}/${p}_${n}_lat_stats/
+            mv ${ERR_PREFIX}* ${OUT_FOLDER}/${p}_${n}_lat_stats/
 	    fi
+	
+
+        #LIBSWING_SEGMENT_SIZE
+        PREALLOC_SIZE=536870912
+        for SEGMENT_SIZE in 1048576 # 0 8388608 1048576 131072 65536
+        do
+            for PORTS in 1 6
+            do
+                MIN_ELEMS=$((PORTS * p))
+                if [ "$n" -ge "$MIN_ELEMS" ]; then
+                    #LIBSWING_DIMENSIONS=${DIMENSIONS} LIBSWING_NUM_PORTS=${PORTS} LIBSWING_ALGO="SWING_B" ${MPIRUN} ${EXTRA_MCAS} ${MPIRUN_MAP_BY_NODE_FLAG} ${MPIEXEC_OUT} -n ${p} ${MPIRUN_ADDITIONAL_FLAGS} ./bench ${COLLECTIVE} INT ${n} ${iterations}
+                    #mv ${OUT_PREFIX}*.0 ${OUT_FOLDER}/${p}_${n}_bw_${PORTS}.csv; rm -f ${OUT_PREFIX}* #${ERR_PREFIX}*
+                    #mkdir ${OUT_FOLDER}/${p}_${n}_bw_${PORTS}_e/
+                    #mv ${ERR_PREFIX}* ${OUT_FOLDER}/${p}_${n}_bw_${PORTS}_e/	
+                    
+                    #LIBSWING_DIMENSIONS=${DIMENSIONS} LIBSWING_NUM_PORTS=${PORTS} LIBSWING_ALGO="SWING_B_COALESCE" ${MPIRUN} ${EXTRA_MCAS} ${MPIRUN_MAP_BY_NODE_FLAG} ${MPIEXEC_OUT} -n ${p} ${MPIRUN_ADDITIONAL_FLAGS} ./bench ${COLLECTIVE} INT ${n} ${iterations}
+                    #mv ${OUT_PREFIX}*.0 ${OUT_FOLDER}/${p}_${n}_bw_coalesce_${PORTS}.csv; rm -f ${OUT_PREFIX}* #${ERR_PREFIX}*
+                    #mkdir ${OUT_FOLDER}/${p}_${n}_bw_coalesce_${PORTS}_e/
+                    #mv ${ERR_PREFIX}* ${OUT_FOLDER}/${p}_${n}_bw_coalesce_${PORTS}_e/
+                    
+                    #LIBSWING_DIMENSIONS=${DIMENSIONS} LIBSWING_NUM_PORTS=${PORTS} LIBSWING_ALGO="SWING_B_CONT" ${MPIRUN} ${EXTRA_MCAS} ${MPIRUN_MAP_BY_NODE_FLAG} ${MPIEXEC_OUT} -n ${p} ${MPIRUN_ADDITIONAL_FLAGS} ./bench ${COLLECTIVE} INT ${n} ${iterations} 
+                    #mv ${OUT_PREFIX}*.0 ${OUT_FOLDER}/${p}_${n}_bw_cont_${PORTS}.csv; rm -f ${OUT_PREFIX}* #${ERR_PREFIX}*
+                    #mkdir ${OUT_FOLDER}/${p}_${n}_bw_cont_${PORTS}_e/
+                    #mv ${ERR_PREFIX}* ${OUT_FOLDER}/${p}_${n}_bw_cont_${PORTS}_e/	
+                    
+                    LIBSWING_DIMENSIONS=${DIMENSIONS} LIBSWING_PREALLOC_SIZE=${PREALLOC_SIZE} LIBSWING_SEGMENT_SIZE=${SEGMENT_SIZE} LIBSWING_NUM_PORTS=${PORTS} LIBSWING_ALGO="SWING_B_UTOFU" ${MPIRUN} ${EXTRA_MCAS} ${MPIRUN_MAP_BY_NODE_FLAG} ${MPIEXEC_OUT} -n ${p} ${MPIRUN_ADDITIONAL_FLAGS} ./bench ${COLLECTIVE} INT ${n} ${iterations} 
+                    mv ${OUT_PREFIX}*.0 ${OUT_FOLDER}/${p}_${n}_bw_utofu_${PORTS}_${SEGMENT_SIZE}.csv; rm -f ${OUT_PREFIX}* #${ERR_PREFIX}*
+                    mkdir ${OUT_FOLDER}/${p}_${n}_bw_utofu_${PORTS}_${SEGMENT_SIZE}_e/
+                    mv ${ERR_PREFIX}* ${OUT_FOLDER}/${p}_${n}_bw_utofu_${PORTS}_${SEGMENT_SIZE}_e/
+                fi
+            done
         done
 		
         echo " ${GREEN}[Done]${NC}"
